UpdaTA system update management, first commit.
We have a tool now which determines how the system in its current
state must be updated to end up in a given state (dnf or recovery,
which versions of what and where to take these data from). The tool
generates a complete plan in JSON format which can be processed by
other tools.

Change-Id: I2ccc5cde9c8731cdec5229c0070b3092d96e96ea