updata_execute: No version check after recovery system upgrade.
We cannot make an exact version comparison because the recovery system
works for a whole range of StrBo versions. The recovery system's
version frequently does not match the target version. Hence, we rely
on updata_determine_strategy to do the right thing, and just log the
recovery system version information after its upgrade.

Change-Id: I421652dac7f1985d2ca7eac531746b35aaf1f0dc