updata_execute: Options for splitting the update into two parts.
In case we don't want the system to reboot immediately after having
prepared all the rest, we can now pass option --avoid-reboot. This
asks updata_execute.py to do everything that is needed according to
the update plan, but it avoids a reboot if such a step was planned.

In contrast, option --reboot-only asks updata_execute.py to do *only*
the reboot in case the plan contains such a step.

Passing none of the options results in full execution of the update
plan. Passing both options avoids all actions altogether, meaning that
nothing will be done.

Change-Id: I4918080e32450c03de838c41999be7fe8886f3ed