update: Add shell script template which wraps updata_execute.py.
The purpose of this script is to support creating a standalone program
which cannot be uninstalled by the package manager during an update.
It can be run from the REST API, from systemd, or any other suitable
to start or resume an update.

To generate a working script, the following patterns must be
substituted: @THE_PLAN@, @STAMP_DIR@, and @ALLOW_EXECUTION@.

Pattern @THE_PLAN@ must be replaced by an absolute filename of an
update plan as generated by updata_determine_strategy.py.

Pattern @STAMP_DIR@ must be replaced by an absolute directory in which
this script can maintain its state files.

Pattern @ALLOW_EXECUTION@ must be replaced by the string "yes". This
is a guard against running the template file directly.

Change-Id: I4a4e431d4f496cc6e0492b3a3c1b222605c9f503