updata_determine_strategy: Fix up dnf flavor variables.
In case we are starting with a fresh image made for some specific
flavor, the dnf variables which contain the flavor information are not
set. On update request for the latest version of the current flavor,
the system tells us that we are already running the flavor, but dnf
will not know about it. Without telling dnf, it will only see the
stable repository, and the update is almost certain to fail.

A possible fix would have been to generate the image with correctly
set dnf variables, but a better fix is to have this discrepancy
detected and fixed at runtime. This will catch not only the case
described above, but also any misconfigurations in this respect.

Change-Id: Ic48c6f5efdf20fa7cc2c903178f5a20ee973b95e