updata_execute: Download packages first, then install.
We also ask systemd to enter the system update target so that all our
software is stopped reliably.

Since the network connection might be dead in that state, we download
all required packages in an explicit step before installing then. Dnf
does this anyway, but we want to download while there is a network
connection, then enter update state, then perform the offline update.

Change-Id: I3105ec68ee4e172739e093d9b3bc8424cd8599c3